LayoutTests/fast/dom/background-shorthand-csstext-expected.txt

This page tests whether or not the background shorthand properly omits initial values.

PASS: document.body.style.background == 'green' should be true and is.
PASS: document.getElementById('div1').style.background == 'repeat-x, white repeat-y' should be true and is.
PASS: document.getElementById('div2').style.background == 'blue 50% 50%' should be true and is.
PASS: document.getElementById('div3').style.background == 'rgb(255, 255, 255) none repeat scroll' should be true and is.
